Add means to define absolute path of app
Running the app as an Airflow task requires that the path to the
app can be defined by the app or user (as a command line arg).
This commit includes both, as the app first finds the path to the
neuropsych_summary_scrape.py file and then if the user supplies
a command line argument to the --app_path option, the app
overwrites the code-determined app path by what the user passed.

The README has also been reörganized and updated.